refactor: split src/index.ts into pure library entry + bin shim
Before: src/index.ts was the CLI entry point. It had a shebang, built
the Commander program at module-load time, exported \`program\` (so the
cli-index test could drive it), AND exported a handful of public API
symbols at the bottom. Importing \`freelance-mcp\` as a library triggered
Commander construction and all CLI handler imports as side effects —
not what anyone doing \`import { createServer } from "freelance-mcp"\`
expects from a module import.

After:

- src/index.ts — pure library re-exports. No shebang, no Commander, no
  side effects on import. Exposes GraphBuilder, GraphEngine, EngineError,
  createServer, startServer, TraversalStore, MemoryStore and the types
  that go with them. Stays at the \`.\` export of package.json.

- src/bin.ts — 4-line shebang entry that imports \`program\` from
  cli/program.js and calls parseAsync. This is the new \`bin.freelance\`
  target in package.json. Compiles to dist/bin.js.

- src/cli/program.ts — the Commander program construction. Exports
  \`program\` so tests can drive it without invoking parseAsync. Mirrors
  the previous src/index.ts content verbatim with paths adjusted for
  the deeper directory (\`../server.js\` instead of \`./server.js\`, etc.)
  and an extra \`..\` in the completion command's templates path.

- package.json: bin → dist/bin.js, dev script → tsx src/bin.ts.

- test/cli-index.test.ts: import program from \`../src/cli/program.js\`.

Eliminates the isMain guard at the bottom of the old src/index.ts —
bin.ts is the only caller of parseAsync, no guard needed.
