+++ src/bin.ts
@@ -1,0 +1,4 @@
+#!/usr/bin/env node
+import { program } from "./cli/program.js";
+
+program.parseAsync(process.argv);
